META-312: Client side renamings: 'PluginEvent' -> 'PluginDiagnosticEvent'
META-312: Fixed a couple of log messages in the stub plugin UT
META-312: Replaced 'PluginEvent' -> 'PluginDiagnosticEvent' in the server and common code
META-315: Added logs to AnalyticsEventReceptor
META-315: Added more analytics logging on the server side
META-315: Minor coding style fixes
META-317: Fix the bug: usage of expression generators leds to bad named directories creation
META-317: move tegra_video_analytics_plugin into dedicated dir
META-318: Add missing copyright notice
META-318: Analytics SDK errors are reported via return values
META-318: Analytics SDK: Code improvements, including error reporting and comments
META-318: Analytics SDK: Improve logging code, fix coding style
META-318: Analytics SDK: Slightly improve code of error reporting
META-318: Fixed a plugin side setting in the stub analytics plugin
META-318: Fixed comments
META-318: Minor code improvements
META-318: Unified SDK error reporting, introduced IError interface
META-324: Analytics SDK: Adde a mutex to RefCountableRegistry - fix races
META-324: Analytics SDK: Improve logging in RefCountableRegistry - always log start and finish
META-324: Analytics SDK: Improved RefCountableRegistry verbose mode: print object addresses
META-324: Analytics SDK: Improved impl of RefCountableRegistry
META-324: Analytics SDK: Initial impl of RefCountableRegistry, with some unresolved todos
META-324: Analytics SDK: Logging improved in RefCountableRegistry
META-324: Analytics SDK: Make Ptr ctor from raw pointer private in favor of makePtr() and toPtr()
META-324: Fix distribution building for bpi and edge1
META-324: Make libs nx_sdk and nx_plugin_utils static for non-Windows as well
META-337: Provided fields 'desc' and 'version' for analytics engines manifests
META-338: Added the isActive flag to the PluginInfo structure
META-338: Minor English grammar fix
META-340: Analytics SDK: fixed a comment
META-340: Analytics SDK: renamings id -> trackId
META-340: Changed analytics database scheme
META-340: Minor renamings
META-340: Renamed some analytics related files
META-340: Renaming: AnalyticsDbObjectCache -> AnalyticsDbObjectTrackCache
META-340: Unified usages of the term 'object track' in the Server and the Client codebase
META-342: Fix linux build
META-342: SDK: In Windows, set and restore DllDirectory for each plugin's entry point function
META-345: Fixed analytics database filter deserialization
META-46: Change compiler, that used to build Analytics SDK, to the minimal required by docs (5.4.0)
META-46: In Analytics SDK CMake file: fix grammatic bug and improve coding style
META-46: Rename Analytics SDK artifacts to meet RDep restriction (artifact name clashing)
VMS-14681: Allow click focus for event panel proxy widget

<==== Mercurial data ====>
Branch: vms_4.0

+++ vms/server/nx_vms_server_db/src/nx/vms/utils/system_merge_processor.cpp
@@ -787,7 +787,7 @@ bool SystemMergeProcessor::fetchUserParams(
     std::vector<std::tuple<QnUuid, ec2::ErrorCode, nx::vms::api::ResourceParamDataList>>
         getUsersParamsResults;
 
-    nx::utils::Counter expectedResponseCount(users.size());
+    nx::utils::Counter expectedResponseCount((int) users.size());
 
     for (const auto& user: users)
     {
